+++ cogs/admin.py
@@ -30,17 +30,19 @@ from textwrap import indent
 from traceback import format_exc
 from utils.utils import localize_time, pairs, run_in_subprocess, generate_activity
 from re import finditer
-from typing import Union
+from typing import Union, Optional
 from dreambot import DreamBot
 from utils.checks import ensure_git_credentials
 from utils.network_utils import network_request, NetworkReturnType
 from utils.database.helpers import execute_query, retrieve_query
 from aiosqlite import Error as aiosqliteError
-from datetime import datetime
+from datetime import datetime, timedelta
 from utils.context import Context
 from copy import copy
 from importlib import reload
 from utils.logging_formatter import bot_logger
+from asyncio import sleep
+from discord.ext import tasks
 import discord
 import sys
 import re
@@ -68,6 +70,7 @@ class Admin(commands.Cog):
 
         self.bot = bot
         self._last_result = None
+        self.logging_line_break.start()
 
     async def cog_check(self, ctx: Context) -> bool:
         """
@@ -141,6 +144,39 @@ class Admin(commands.Cog):
         except commands.ExtensionNotLoaded:
             await self.bot.load_extension('cogs.' + module)
             await ctx.send(f'Loaded Module: `{module}`')
+
+    @commands.command(name='sync')
+    async def sync_commands(self, ctx: Context, sync_type: Optional[str] = None) -> None:
+        """
+        Syncs application commands based on the specified sync type.
+        Defaults to syncing to the local guild.
+
+        Checks:
+            is_owner(): Whether the invoking user is the bot's owner.
+
+        Parameters:
+            ctx (Context): The invocation context.
+            sync_type (Optional[str]): The type of sync to perform.
+
+        Returns:
+            None.
+        """
+
+        if sync_type == 'global':
+            synced = await ctx.bot.tree.sync()
+        elif sync_type == 'local':
+            synced = await self.bot.tree.sync(guild=ctx.guild)
+        elif sync_type == 'clear':
+            ctx.bot.tree.clear_commands(guild=ctx.guild)
+            await ctx.bot.tree.sync(guild=ctx.guild)
+            synced = []
+        else:
+            ctx.bot.tree.copy_global_to(guild=ctx.guild)
+            synced = await ctx.bot.tree.sync(guild=ctx.guild)
+            sync_type = 'from global'
+
+        await ctx.send(f'Synced {len(synced)} commands using sync type: `{sync_type}`.')
+        bot_logger.info(f'Synced {len(synced)} commands using sync type: {sync_type}.')
 
     @commands.command(name='unload', hidden=True)
     async def unload(self, ctx: Context, module: str) -> None:
@@ -600,6 +636,35 @@ class Admin(commands.Cog):
         embed.set_footer(text="Please report any issues to my owner!")
 
         await ctx.send(embed=embed)
+
+    @tasks.loop(hours=24)
+    async def logging_line_break(self) -> None:
+        """
+        Inserts a logging line break at the start of each day.
+
+        Parameters:
+            None.
+
+        Returns:
+            None.
+        """
+
+        bot_logger.info(f'----- Line Break Inserted for Readability -----')
+
+    @logging_line_break.before_loop
+    async def before_logging_line_break(self) -> None:
+        """
+        A pre-task method to ensure the bot is ready before executing.
+        Additionally sleeps until the start of the next day before inserting the first line break.
+
+        Returns:
+            None.
+        """
+
+        await self.bot.wait_until_ready()
+        today, now = datetime.today(), datetime.now()
+        time_until_tomorrow = (datetime(year=today.year, month=today.month, day=today.day) + timedelta(days=1)) - now
+        await sleep(time_until_tomorrow.total_seconds())
 
     @commands.command(name='as', hidden=True)
     async def execute_command_as(