+++ dreambot.py
@@ -35,6 +35,8 @@ from copy import deepcopy
 from utils.logging_formatter import bot_logger
 from uuid import uuid4
 from json import dumps
+from sys import stderr
+from traceback import print_exception
 import discord
 
 
@@ -119,8 +121,9 @@ class DreamBot(Bot):
             if cog.endswith('.py') and cog[:-3] not in self.disabled_cogs:
                 try:
                     await self.load_extension(f'cogs.{cog[:-3]}')
-                except ExtensionError as e:
-                    bot_logger.error(f'Failed Setup for Cog: {cog[:-3].capitalize()}. {e}')
+                except ExtensionError as error:
+                    bot_logger.error(f'Failed Setup for Cog: {cog[:-3].capitalize()}. {error}')
+                    print_exception(type(error), error, error.__traceback__, file=stderr)
 
     @tasks.loop(minutes=30)
     async def refresh_presence(self) -> None:
@@ -226,8 +229,8 @@ class DreamBot(Bot):
                 'INSERT INTO BASE_EVENTS (ID, DATE, EVENT, DATA) VALUES (?, ?, ?, ?)',
                 (event_id, event_timestamp, event, event_data)
             )
-        except aiosqliteError:
-            pass
+        except aiosqliteError as e:
+            bot_logger.error(f'Failed to Record Base Event. Error: {e}. Params: {event}, {data}')
 
 
 async def get_prefix(bot: DreamBot, message: discord.Message) -> List[str]: